[SYSTEMML-1236] Move MLContextProxy out of api package
MLContextProxy is used internally by SystemML but not my end users, so move
this class to a non-API package.
Remove some unnecessary casting since only 1 MLContext class now exists.
